Make CI green: pin the lint contract, and stop skipping bundle verification
Every CI run since the first commit has failed — 9 for 9, the badge in the
README red the whole time. None of them failed on a test. All nine died on
`ruff check src tests`, which runs before pytest, so the test job never
executed at all on any commit in this project's history.

The cause is a floating floor. `ruff>=0.5` resolves to whatever ruff released
most recently, and ruff has since moved a large number of rules into its
default set that were not in it when that floor was written. Nothing in this
repository changed to become non-compliant; the tool's definition of
compliance moved underneath it. Today that was 69 findings, every one a style
or modernisation rule and not one a correctness rule.

So the fix is the contract, not just the findings: `[tool.ruff.lint].select`
now spells the rule set out and the version is bounded to `>=0.16,<0.17`. A
new ruff release can no longer turn this repository red on its own. Two rules
are ignored with the reason recorded next to them — TRY003, because naming the
offending value at the `raise` site is what makes a failed run diagnosable,
and B905, because `zip(strict=)` changes behaviour rather than style and does
not belong in a lint cleanup.

Under that set the tree is clean. Most of it is mechanical (import order,
`collections.abc`, unquoted annotations under `from __future__ import
annotations`, `int(round(x))` where `round` already returns an int). Four were
worth having: a leaked file handle in the row count printed by `cli submit`, a
`SyntheticCostModel()` built once at import time and shared by every dry run,
a dead `for ... : pass` loop, and 18 implicit string concatenations inside
audit.py's report list, where a missing comma would have silently merged two
report lines.

Verified inert rather than assumed to be. `audit.py` parses to a byte-identical
AST. The full synthetic pipeline — labels, sampling, features, B3, CQR,
decision, regret-k, ALNS, routing, cost — and the grouped CV harness both
produce output identical to the previous commit's, character for character.
The vendored modules were re-parsed against Python 3.10, the container's
version, and all of them carry `from __future__ import annotations`, so the
unquoted annotations stay lazy there.

Two bundle tests were also failing on any fresh checkout, and would have kept
CI red after the lint fix. `verify_bundle` needs `script.py` and
`requirements.txt`, which come from the official example repo — cloned, not
vendored, and git-ignored, so no CI runner has ever had it. Their sibling
guards itself with a skipif; these two did not. Skipping them was the obvious
move and the wrong one: it would mean the bundle verification never runs in CI
at all, which is exactly the check worth having before a submission. Instead a
fixture stands in for the three files, built from the runtime package set the
competition fixes, and the real clone is preferred when present. The check now
runs everywhere and still reads the genuine files locally.

92 passed, 9 skipped, ruff clean, and all three CLI steps CI runs exit 0.

+++ src/batteryswap/pipeline.py
@@ -181,7 +181,7 @@ def train(X: pd.DataFrame, y: pd.Series, groups: pd.Series, seed: int,
           calib_fraction: float = 0.25) -> TrainedModel:
     """Fit B3 and conformalise it on held-out BUILDINGS (never rows)."""
     buildings = sorted(groups.unique())
-    n_calib = max(1, int(round(calib_fraction * len(buildings))))
+    n_calib = max(1, round(calib_fraction * len(buildings)))
     rng = np.random.default_rng(seed)
     calib_buildings = set(rng.choice(buildings, size=n_calib, replace=False))
 
@@ -198,7 +198,7 @@ def train(X: pd.DataFrame, y: pd.Series, groups: pd.Series, seed: int,
 
     y_arr = y_cal.to_numpy(dtype=float)
     diagnostics = {
-        "n_fit": int(len(X_fit)), "n_calib": int(len(X_cal)),
+        "n_fit": len(X_fit), "n_calib": len(X_cal),
         "fit_buildings": int(len(buildings) - len(calib_buildings)),
         "calib_buildings": sorted(calib_buildings),
         "mae_days_raw": mae(raw[q_col(0.50)].to_numpy(), y_arr),
